+++ sidekick_core/lib/src/template/entrypoint.template.dart
@@ -1,3 +1,9 @@
+String entrypointTemplate({required String packagePath}) {
+  return _code.replaceAll('{{packagePath}}', packagePath);
+}
+
+// language=Bash
+const String _code = r'''
 #!/usr/bin/env bash
 set -e
 
@@ -19,4 +25,5 @@ cd "$(dirname "$PRG")/" >/dev/null
 export SIDEKICK_ENTRYPOINT_HOME="$(pwd -P)"
 cd "$SAVED" >/dev/null
 
-"${SIDEKICK_ENTRYPOINT_HOME}/{{{packagePath}}}/tool/run.sh" "$@"
+"${SIDEKICK_ENTRYPOINT_HOME}/{{packagePath}}/tool/run.sh" "$@"
+''';